Bug Fixes + Improvements
Digging Minigame saves distance travelled and passes it back to player in game view so the game can continue. The player no longer spawns on the right of the screen. The player dies when going down out the screen. The player can no longer be thrown off screen when being attached